Changing predicate column type

+++ reworking/src/helpers/predicate.cpp
@@ -2,15 +2,16 @@
 #define PREDICATE
 
 #include <cstdint>
+#include <cstddef>
 
 class Predicate
 {
 public:
     float low;
     float high;
-    float column;
+    size_t column;
 
-    Predicate(float low, float high, float column)
+    Predicate(float low, float high, size_t column)
         : low(low), high(high), column(column){}
     ~Predicate(){}
 };
